requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.23
pydantic==2.5.0
sqlalchemy==2.0.23
redis==5.0.1
rq==1.15.1
openai==1.3.7
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
psycopg2-binary==2.9.9
alembic==1.13.0 